+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 - Integration tests https://github.com/xcodeswift/xcproj/pull/168 by @pepibumur
 
+### Fixed
+- `PBXGroup` not generating the comment properly for its children https://github.com/xcodeswift/xcproj/pull/169 by @pepibumur.
+
 ## 1.5.0
 
 ### Added