+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Integration tests https://github.com/xcodeswift/xcproj/pull/168 by @pepibumur
+- More examples to the README https://github.com/xcodeswift/xcproj/pull/116 by @pepibumur.
 
 ## 1.5.0